[RPP] Introduce SyntheticEvents helper
Its purpose is to create and track all synthetic events made in the
trace engine. The reason being, we need to easily access made synthetic
events when deserializing event keys from modifications in a trace file.
At the moment this includes synthetic events created from
a raw trace events. Support for Profile calls will be added as a
follow up.

The helper is supposed to be called for every synthetic event to be
created, to ensure that all synthetic events are held in an accessible
place. To help enforce this, a branded type is introduced for synthetic
events. The helper receives a trace event and returns a synthetic event
with the branded type. The idea is that if a synthetic event is created
w/o the helper, a forced cast needs to be used to meet the branded
type, which would make the faulty synthetic event creation more evident.

So that synthetic events can be found easily from a key, they are stored
in an array indexed based on the position of the corresponding raw event
in the rawEvent array.

Because the helper is called by the trace handlers, it needs to be
set up before handlers are (and a new instance is created for each
trace).

Bug: 339163152
Change-Id: If88ee1d947e50890ab36703bf0d175b9acff0fe6