RPP: fix over-eager level skipping
This CL fixes a bug where the moment an entry level is 1px off the top
of the screen, we skip rendering it. This causes a jumpy experience as
the user scrolls, as events suddenly disappear even though they are
partially visible within the panel.

The fix here is to check not the offset of the level (which is its top y
value), but its offset + height (bottom y value), which represents if
the entry is fully off-screen or not.

Bug: none
Change-Id: I82fec4cdef649311d04691f31e8c2ef087ca73b8