Remove "(async)" suffix for Async Stack Frames
Fixed: 343750870

I removed the "(async)" suffix from Async Stack descriptions.

To align the visuals with the Call Stack panel I switched the Stack trace to use bold instead of italic to indicate an async description. This makes it more clear that these rows are special and not just another stack frame.

Change-Id: Ib113692d23e45468a1ae302591cd881af6dc33f7