[RPP] Refactor how SyntheticEventsManager is initialized
By making part of a trace engine model instance in order to simplify the new implementation of the ModificationsManager in an upcoming CL.

Also, small drive-by changes to simplify further the impl: store
profile calls in the exported data of the renderer handler.

Note: no functional or UI changes, only code refactorings.

Bug: 339163152
Change-Id: I2fad7fff4cefb85b771ee4bbd8b511800b4a662d