Allow console messages containing objects to be seen as equal
When comparing console messages for equality, allow console messages
which contain objects with the same `objectId` and which were sent at
the same timestamp, to be seen as being equal.

This fixes the behaviour of showing multiple identical console
messages when navigating using the back/forward-cache.

Screenshots: https://imgur.com/a/fajNcTD

Bug: 40894153
Change-Id: I8daab02a9a4dd0e510cba2ce6b884b255b641533