RPP: Create overlay for Entry Selected
This CL creates an overlay for when the user has selected an entry on
the timeline and replaces the existing one.

Note that this only happens if the user has enabled the experiment for
svaing traces with modifications. This is a slightly odd experiment to
use, but given that we hope to push this into stable very soon, this is
a temporary choice to allow the team to test this in Canary before we
land it. It is not expected that users will enable this.

This CL has one caveat which is that users are unable to select frames
in the frames track, as they are not represented as trace events. I have
logged this in b/344771436 and will work on this in a follow-up.

Bug: 343388196, 344771237
Change-Id: I6fc916ca9534b5038706be1e0bd31459b9e6abb1