An explicit paramter to constrol throttler scheduling behavior.
Currently throttler schedules the task immediately if it doesn't have
pending tasks or if the second paramter is true.

This default behavior is often what we want for throttling heavy
operatins but is not suitable for debouncing. It is also somewaht
suprising that setting asSoonAsPossible to false can still schedule
immedimmediately.

This will allow to simplify VE logging implementation and test, where to
ensure debouncing we currentlt schedule an empty task before scheduling
the real task, so that the latter will never run immediately.

Bug: 343845140
Change-Id: If5888ed491ef3d9693f2df714e8e2c06e33e2ff1