Hide ignore-listed frames from ConsoleViewMessages
This follows the approach already in JSPresentationUtils which deals with hiding frames in the expanded stack trace of the log entry.
This does the same thing for the linkified message itself.

One thing that's different here is that the message is likely visible from the start so it's much more likely that this will cause a flash and relayout after the source map loads asynchronously.
However, that's already the case that the file names themselves change in size and sometimes cause relayout anyway. This just makes it more likely that an error that happens during page load flashes a bit as souce maps load.

Currently, I don't include a way to expand the frames because it's cleaner when the message can remain unbothered with that and I'm not sure how you want to treat that UI. However, I find that it's not required here anyway because typically in a reported error, you'd have the stack in the expanded little tringle anyway and that lets you dig deeper. The original error.stack isn't very useful for digging deep anyway since by default it bottoms out at 10 frames where as the expanded one will have the full stack.

Bug: 345248263
Change-Id: I0d4e0b2eae57bb0bb8bc8fa5bf2c004568f1d125