RPP: use isInPrimaryMainFrame to identify main frame
This CL makes the frontend use the isInPrimaryMainFrame flag added to
the TracingStartedInBrowser event in crrev.com/c/5595033 in order to
very accurately identify which frame is the main frame. This is
superior to the existing method if using isOutermostMainFrame because
multiple frames can have this flag set to true (because of MPArch,
pre-rendering, and so on).

Fixed: 343873756
Change-Id: I6bb5a0a3d191936912e55dbd9d656978f65c1eb8