Fix Network overview grid resizer click
Issue:
In Network Overview Grid, the resizers does not have focus when clicked, so users cannot click and use left/right arrow key to replace hold and drag, and when using the Voice Access on Windows, users cannot select the resizer and then resize with voice command.

Fix: Assigned focus when the resizers clicked.

Fixed: 343942719
Change-Id: I5a2afaed2fae1df2bd0ae35ab1f9c2245ae07f97