[l10n] Fetch en-US.json only once (or thirce) for unit tests
We don't need to refetch en-US.json for each individual test if the
strings are already there.

The only tricky bit is that we have unit tests for
'fetchAndRegisterLocaleData' that stub out `fetch` and check that we
fetch from the right URL based on whether a locale is bundled or not.
These tests leave the locale data in a broken state so we need to
clear it up after the i18n unit tests.

This means that now we fetch en-US.json at most 3 times.

R=bmeurer@chromium.org

Bug: None
Change-Id: I42ac7e64d7a00e092a515a096c35895badeae51e