[RPP] Switch to correct Modifications Manager when switching between traces
We had a bug that caused the modifications manager to always be set to the one for the latest loaded trace because the  index of the active trace was not updated when switching between traces.
This resulted in trace data not displaying after switching traces because a breadcrumb from the last loaded trace was applied to it.

Now we call initAndActivateModificationsManager when setModel is called, which happens both during a new trace load or switching between traces. In initAndActivateModificationsManager, we assign active manager to a new or existing one.

Bug: 345627609
Change-Id: If7f699a95fa6c26153e6129fef078e5e9e0062c3