RPP: Zoom to activity of busiest main-thread track
In some traces, we were setting the zoom window based a secondary thread, leaving the user zoomed in on an irrelevant time-range. We fix this in two pieces:

The sort order of main threads and frames is updated. Once about: and chrome: URLs are deprioritized, the remaining ones are sorted by entries count, descending. This ensures that if there were navigations in the trace, the top-most track will be the busiest, and likely the intended focus.

Our auto-zoom to main-thread-activity feature consumed thread data, but
instead we can align the top-most track with the entries chosen for autozoom calculation.

Bug:345599356
Change-Id: I6c9408c9ba2d1027d5f797d0a1b9afeb49c9711c