Add time range overlay
This CL adds the fundamentals of time range overlays behind an
experiment. If the experiment is enabled, shift+click & dragging will
show the time range using the new overlays system.

There is no way currently to make an annotation time range - that work
will come later as part of the annotations epic.

Time ranges can support a plain text label too. As part of annotations
this will need to be editable, but that is also not implemented yet.

Follow-up work is:
- creating & editing these (done as part of annotations)
- dealing with ranges overlapping and how they look (follow-up CL)
- work to optimise positioning and skip if they are out of view
  (follow-up CL)

Bug: 345411951
Change-Id: Ibbbe7cfb3c807cc076fd750acf22c29745c8f2ac