RPP: fix overlay positioning with collapsed network
This CL fixes a bug where if the network track was collapsed, overlays
would be incorrectly positioned on the y axis. This was because the
positioning logic took into account the network height + the resize bar,
but when the track is collapsed there is no resize bar.

The CL fixes it by having the network track report if it is collapsed,
and then the overlay positioning can take that into account.

Bug: 343388196
Change-Id: I24a286930308d377359af209b1929be5b2b24201