Add `formatMicrosecondsTime` to i18n utils
We have time formatters in the trace engine, but they are different to
the `preciseMillisToString` method in i18n. We should use that
formatting as much as possible, so this CL adds a
`formatMicrosecondsTime` method to the i18n module which we will then
use across the panel, and remove the trace engine specific helpers (it
also doesn't really make sense for a parsing engine to ship formatting
helpers).

Bug: 346576296
Change-Id: I8ef01b0c353e16de44f63209793f11d5361a0330