RPP: remove metric score property
The score property is just the time formatted as seconds - which we can
do via the i18n helper. So this CL removes the score from the handler
(which shouldn't care about formatting numbers) and leaves it up to the
UI.

Because the formatting helpers in the trace engine are going to be
removed (follow-up CL), this also adds in a method to format a time and
force it to be done in seconds, which is what we like to do for many CWV
scores. This method can be used in place of using the `score` property
that the handler was setting.

Bug: 346576296
Change-Id: I30b590ea876483c82553dcd0721c4c074253bd22