RPP: turn selected entry overlay on
This CL removes the old logic for highlighting selected entries and
moves it over to the new overlays system by default.

Because some non-perf-panel flame charts unfortunatley rely on the old
selected entry system, this CL also includes a mechanism to disable it.

It also fixes a bug where selected entries would be visible even if they
were in a collapsed group.

Bug: 343388196
Change-Id: Id1f82aa7a6c19544a09b95752b7ad87b886e4b01