[RPP Observations] Refresh rendered nodes on document update
If a document update happens after we render the related node links then
the related node links won't work because there is a new forntend node
ID associated with each node.

This CL "refreshes" each node and sends an update to the UI whenever
a document update occurs.

This CL also removes the related node from the INP event since we
aren't currently using it.

Bug: 313906462
Change-Id: I9d8ae0eb7804769002919c30dd7c573cee72e3ee