Added default case for arrowFunctions in jsFormatter
javascriptFormatter.js does not have a set spacing for arrow functions, In turn the formatter defaults to no spaces / removes any current spacings.

Added default spacing on both sides of an arrow function => when using the javascriptFormatter {} in devtools>sources>snippets
Wrote tests to confirm behavior is as expected.

R=szuend@chromium.org

Fixed: 339172103
Change-Id: Ia3f5c08312344b84bf3d760b62fdd78e55e802aa