Cache results of checking for WeakMap edges in heap snapshot
DevTools uses a regular expression to check whether edges in a heap
snapshot represent part of a pair by which a WeakMap and its key keep
the corresponding value alive. This takes some time, and most edges do
not match that regular expression, so we can speed up the process by
keeping a bitmap of which strings are known to not match the regular
expression. On a heap snapshot from Google Docs, this change speeds up
HeapSnapshot.initialize from about 3.2 seconds to 2.9 seconds.

Bug: 40818377
Change-Id: I832849be0db9b7099b2b25df274b182a6a19ba16