[trace engine] Add Lighthouse's Lantern library
This is the culmination of isolating Lighthouse's Lantern library
(https://github.com/GoogleChrome/lighthouse/issues/15841). The code in
this CL took Lantern as it existed in Lighthouse, converted to
TypeScript, integrated its unit tests (Jest -> Karma conversion), and
resolved the majority of CDT's JavaScript linter errors.

There is no behavior change to CDT with this CL, besides TraceModel
exporting the Lantern module.

A handful of eslint checks have been disabled for this folder, some or
all of them will be addressed in later CLs.

Bug: 335099340
Change-Id: Ib2bf64fc17d334859caca5c42bea1e276465d64c