RPP: log selected group when entry is selected
This CL implements some logging for the groups/tracks on the performance
panel:

1. when a trace is recorded/loaded, impressions for each group are
   logged
2. when the user selects an entry, we log a click and log the group's
   name as its context

To ensure we do not log any PII, we don't log the full group name for a
track (which may include a URL) but instead log from a list of options
representing the types of tracks we have and how we want to log them.
This is similar but distinct from our list of track names, as we want to
be more specific than "Thread" and identify main thread, other frames,
workers, etc, etc.

Bug: 344852406
Change-Id: I56d34dd6bad3bae85a6002e640242bb33527e6f4