[sdk] Use the source map 'scopes' information to calculate inlining info
This CL adds a new "SourceMapScopesInfo" object that will contain
various heplers to calculate scopes related information.

First up is inlining: This CL adds a method that, given a generated
position, will find all the original function names inlined into the
surrounding generated function. We'll use this to expand call frames
on debugger pause similar to what the DebuggerLanguagePlugin is
doing for DWARF.

R=kimanh@chromium.org

Bug: 40277685
Change-Id: I448fd9b5ed50f778cccda4e7b01a114389aa48d1