[RPP] Create LCP phases overlay
The overlay is triggered by a click of the insight in the
sidebar.

TimelineFlameChartView.ts constructs the overlay and calls
Overlay.ts which handles how the overlay is defined and positioned.

TimespanBreakdownOverlay.ts largely borrows from TimeRangeOverlay.ts,
with specifics for the TimespanBreakdownOverlay for handling
the breakdown's labels etc.

This also fixes a little bug in LargestContentfulPaint.ts.

Bug: 344550388
Change-Id: Id8fd0d31a943495f63b8555bba90ecac340a03f4