Speed up computation of dominators for heap snapshots
Computation of dominators spends a lot of its time iterating over a very
sparse array to find which nodes should be visited. We can speed up that
iteration by packing the dirtiness information into a single bit per
node. On a test snapshot, this change decreases the time spent in
HeapSnapshot.buildDominatorTree from 1.1 seconds to 760 milliseconds.

Bug: 42203857
Change-Id: I4103c10c191b3115d3871254aa3b52225ad0e4ec