[console] Remove unnecessary 'newline' when rendering Errors
The CL that added Error.cause support (https://crrev.com/c/5385648)
changed the formatted error HTML element from 'span' to 'div'. This
causes uncaught exception to be rendered with an unnecessar newline
between 'Uncaught' and the actual error:

```
Uncaught
Error: Test
    at <anonymous>:1:7
```

This CL fixes this by partially reverting the Error.cause CL: We
render the first element as a 'span', but all the nested 'Caused by:'
errors we render in a 'div'.

R=kimanh@chromium.org

Fixed: 348869920
Change-Id: I94d2482520989686081b544809b844fa7329b86c