[RPP] Move Annotations Overlays (for now just EntryLabel) to ModificationsManager
Apart from the Overlays class, we will need the list of user created annotations accessible from the sidebar, as well as from the ModificationsManager when saving and loading them to/from trace file.
Since all of the other saved annotations (atm Breadcrumbs and hidden entries) are stored in and modified through ModificationsManager, let's also move the Overlays that correspond to user-created annotations, to ModificationsManager.

Since ModificationsManager is a singleton, we can easily access the list of annotations form the sidebar to display them (next CL), as well as from Overlays to render them.

Bug: 349530077
Change-Id: Icbdf6dee17b01c07877e403b0c338b07d1e73d3c