[RPP] Display Entry Labels in the sidebar Annotations Tab
Display the Entry Name and Entry Label in the sidebar. The Label is updated as it is typed/added/removed in the timeline.

Video - http://screencast/cast/NTU5NTc2NjUzNzA2MDM1Mnw2NmU1NTVmNC04OQ

To achieve this, when AnnotationModifiedEvent is dispatched from the ModificationsManager, along with updating overlays, pass the list of the updated AnnotationOverlays to the Sidebar and rerender the SidebarAnnotationsTab content with the new overlays.

Bug: 350657815
Change-Id: I9b997bebabdb2f50476445765b889c33337847d7