[RPP Observations] Display field p75 prominently and label histogram
This CL also determines the metric rating at the UI level so that we
can use the same rating logic for both local and field data. In the
future we should consolidate this logic further with the trace engine.

The histogram will eventually go in a tooltip, but to reduce complexity
in this CL the tooltip has been deferred to another CL.

https://screenshot.googleplex.com/BKy7jNUwgyfqswb

Bug: 313906362
Change-Id: I8f6ac410570aec9113e1555d74214d693c9b8184