[Freestyler] Add a button for "Fix this issue" workflow
Disclaimer: we show this button for all the answers in the last message.
I think we should show this only when there is a bug to fix or a suggestion from the LLM. Added a comment for this and we can discuss how to go about it in the future.

See the bug for the screenshot.

Fixed: 350672340
Change-Id: I43caf47754915eea653ec4138e463778e05658e3