Fix scroll margin in Network Throttling settings page
The Network Throttling settings page doesn't have scroll bars. When
tabbing through the input fields of the throttling profiles, and
back-tabbing using (Alt+Tab), the settings view would not correctly
scroll back and thus hide parts of the title. This adds a scroll
margin to the input fields to adjust the scrolling.

Before:
After:
Fixed: 349529648
Change-Id: I48bf89b1e8d1aac17884ecb3718dcf4c9331c5ba