[GM3Restyling] Increase issues line height in issues
With the Buttons migration, the "Unhide all" button in the IssuesPane
gets cut off when the focus is visible. In order to accomodate this,
this CL increases the height of the line and adjusts the margins.

Before: https://i.imgur.com/rmP1VLG.png
After: https://i.imgur.com/8ImBw8d.png
Fixed: 350628672
Change-Id: I0e6eabb76ffa59fab15de0bc887165cda937c504