Move RPP sidebar to floating button
This CL implements the floating button, ensuring that by default when
the sidebar is closed it does not take up any space. When the button is
clicked the sidebar then expands.

Sidebar closed: https://screenshot.googleplex.com/644iQQPsTZ8giyj

Sidebar open: https://screenshot.googleplex.com/4R2B6dkWvHc56Tm

This CL also updates the sidebar to be non-resizable. It does not
include the work to add the annotations count as a small circle on the
button, we will do that as we have built out all the annotations.

Bug: 348590670
Change-Id: Ie6e68748b27b1a0d8df72e16b988b57813eaa6c4
Fixed: 348572151