[Flame Chart] Fix the bug that the initiator arrows are different width
Before we set the |lineWidth| after drawing the arrow line, and since we don't restore the canvas, this will apply to the next arrow. And for the first arrow, it uses the default width, which is thicker.
This change moves the line width setting before drawing, so it fixes the problem

Bug: none
Change-Id: Ib6ac7dd5fa7eceb42058ca57e9328356172ce852