RPP: draw cursor line in the overlays
This CL moves the vertical line drawn when holding shift to the new
overlays system. It does not put this behind a flag, as I think it is an
improvement on the experience for all users.

Now, the line will extend across the entire UI, rather than across only
the canvas you hover over. This makes it much easier to line things up
when debugging a trace.

Bug: 348619191
Change-Id: I74ea7ba3ac4a7309de1378e4456644e83790add8