RPP: limit file imports to .json or .gz
This CL limits the types of files that can be imported into the
performance panel to `.json` or `.gz`. Note that the `accept` parameter
is best effort - it doesn't actually validate anything, so we could
still get bad input files, but this at least prevents users mistakenly
importing the wrong thing via a misclick.

Bug: 349864878
Change-Id: I11304de5ae6a7b886464d7427cd4d454b172324c