[RPP] Add support for track groups in extension tracks
We received the feature request to allow extension data to be
grouped into "sub tracks". Put in another way, a way to group extension tracks.

Because we have such concept with the "native" tracks already (for example with thread pool worker [1]), we can easily implement this for extension tracks.

To do this, a new field called `trackGroup` is introduced to the devtools object of extension entries. On top of using the existing `track` property, the entries are grouped by `trackGroup` as well. The data with this shape is then passed to the extension track appender, which appends the groups to the flamechart appropriately.

[1] https://imgur.com/a/3IgJeSo

Bug: 347681939
Change-Id: I58d636fb099f8c85f728ef660dd40371584f136e