RPP: mark overlays as singletons
Some of the overlays we will have are singletons - we only ever expect
there to be one of them. This CL opens up the ability to denote that in
code and ensure that if you call `add()` and one already exists then we
will just update the existing one. This means callers don't have to
worry about managing this state + checking for the existence themselves.

Bug: 343388196
Change-Id: Ib090ce66899613a48a74cfdb8527b5b330746513