RPP: make EntrySelected overlays singleton
This CL updates EntrySelected to be a singleton, which means that the
overlays system will only ever create one of them. If another is added,
it will replace the first. Also, "replace" here means updating the
existing overlay, which has the bonus of the overlay system not
destroying and then recreating DOM; the element for the overlay is
reused, which helps with performance.

I also took the opportunity to tidy up some types rather than define a
set of strings, taking advantage of TS assertion functions to make the
code clearer.

Bug: 343388196
Change-Id: I6eed1c345983fd1740496bf6b6dc5429870b7266