Grey out inactive position-try sections
Currently we will reuse the simple UX from the '.no-affect' CSS class,
which happens e.g. when you change a matched selector to an unmatched
selector in a regular StyleRule section. This should be a net plus to
position-try UX, while we think about what to do with section titles.

Bug: 40246493
Change-Id: I6c1b62121d6bc52669e7b827b4a064362a54ec72