[deps] Add missing deps for unit test fixtures
The lantern GN unit test targets are missing a dependency on the test
fixtures. This can cause the trace.gz files to not be copied for
incremental builds.

This CL adds the dependency as a public_dep. We can't use "dep" as
ts_library will look for a -tsconfig.json otherwise, which doesn't
exist for `copy_to_gen` targets.

R=alexrudenko@chromium.org

Bug: None
Change-Id: I8b82cb468c3ac325ef1dabb653a9a209402016e3