[deps] Update TypeScript to 5.5.3
The CL also updates @types/karma-chai-sinon to 0.1.20. This is
required to fix the global `sinon` namespace object that our test
code assumes to be there.

We also need to allow access of UMD globals without import in test
code, otherwise we'd have to add an import of sinon to each
individual test file.

DISABLE_THIRD_PARTY_CHECK=Requires a ts_library change for test modules
R=alexrudenko@chromium.org

Bug: None
Change-Id: I449be485a9e3fe03ccc68fad44ee3573a98b3413