RPP: move sidebar to use details
This CL updates the sidebar to use details/summary elements for multiple
navigations which is the most semantic representation.

It also adds code to only allow one to be expanded at any moment in
time.

This CL does remove the rendering of LCP by phase, but separately @aixba
is working on restructuring those to be able to be rendered as
components within the `devtools-single-navigation` component, so those
will be added back shortly in a follow-up CL.

Bug: 348591291
Change-Id: Ibcb233bc7635c6a0fafce9212ea474e7ef5baa72