RPP: recover from importing invalid files
This CL fixes a bug where if you import an invalid trace file after
importing/recording a valid one, you would lose all history.

The panel was calling `clear()` if there was an error, but that would
completely reset the panel even though the previously imported trace was
still present and valid.

This CL fixes that by checking on an invalid import if we already have
traces in the model; if we do we revert back to the most recent one.

Bug: 349864878
Change-Id: I9ccf5e4c6e604921090e4a63de81eb3fc15db62e