[RPP Observations] Move histogram into tooltip
https://screenshot.googleplex.com/C2xH4ubCEwuQEbH

This uses the web componenet dialog element with a defined origin to
create a hover tooltip that contains the field data histogram. The
histogram is always visible and will show "-" for all percentages if
there is no field data.

This CL also adds a `MetricCard` component which makes it easier to
manager the tooltip state of each metric. In the future, I would like
to do more content rendering in the `MetricCard` component iteself, but
for now it is sufficient to pipe everything though `<slot>` elements.

Bug: 313906362
Change-Id: I2e421c8c29f475e60bb678fef17cbaa825f4e16a