Don't emit 'ScopeChainUpdated' event after ScopeChainModel is disposed
This CL makes ScopeChainModel more consistent: It should not emit
update events after it was disposed.

Using this invariant we can simplify the listener method in the
'ScopeChainSidebarPane'. We no longer need to check if we are still
paused. If we were no longer paused, the "call frame flavor" would
have changed to null, which would cause the model to be disposed
and no update event to arrive.

R=pfaffe@chromium.org

Bug: 40277685
Change-Id: Iddd3cd5517e646b012c31370884d96b4a075198b