[scopes] Allow multiple top-level generated ranges
This was discussed in meetings and the source map hackathon. The spec
already implicitly allows this, we just did not implement it.

A common use case is concatanation, where multiple source files are
concatenated into a bundle. In that case we would have multiple
top-level generated ranges that correspond to the top-level scope in
each original source file.

Implementation-wise we add a pseudo root range to the range stack
and grab it's children once we are done decoding.

R=pfaffe@chromium.org

Bug: 40277685
Change-Id: I16a61a80cd79d64624a94c1e77600515d9e65d60